CORRAIDZ-258 draid: use DVA grid bits for mirror
Use the highest bit of the reserved DVA grid bits for mirrored DVA
on a dRAID vdev. This approach is more robust because it does not
depend on metaslab info to determine if a DVA is mirrored or not.

This patch is not backward compatible - it will not work on pools
created without this patch.

Change-Id: I580b8033596176873a862753757fc729934f00d6
Signed-off-by: Isaac Huang <he.huang@intel.com>
Reviewed-on: https://review.whamcloud.com/26131
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Don Brady <don.brady@intel.com>