CORRAIDZ-175 ZFS-TESTS: Modify Test Suite to include dstat
As part of the effort to provide an analysis capability for the performance
regression test suite dstat is added as a first step to provide the information
for performance analsys. In future phases this will replace iostat, mpstat, and
other tools for gathering information. The patch also changes fio to output to
a json file in addition to its redirected output and renames the output files
to make them easier to parse the embedded components in the path.

Change-Id: I4b0494b0da555ab18f9f0974ad2bf2841e3a526e
Signed-Off-By: David Quigley <david.quigley@intel.com>
Reviewed-on: https://review.whamcloud.com/26679
Tested-by: John Salinas <John.Salinas@intel.com>
Reviewed-by: John Salinas <John.Salinas@intel.com>