CORRAIDZ-383 draid: limit rebuild by memory overhead
Current dRAID limits concurrent rebuild by total number of segments.
This works poorly when a metaslab is fragmented with many small
segments, which is often the case for a mirrored metaslab.

A better way is to limit concurrent rebuild by total memory overhead.
This would allow more concurrent segments to be rebuilt when segments
are small yet still limit the total memory overhead of the rebuild
process.

Change-Id: I73a1a73f43a1baddeaa4389ee6d5bd356e044f13
Signed-off-by: Isaac Huang <he.huang@intel.com>
Reviewed-on: https://review.whamcloud.com/29008
Tested-by: Maloo <hpdd-maloo@intel.com>